update level order
