+++ dojo-dungeon/src/dojo/rooms/level_2/L2_R3_Monster_Implement_2.java
@@ -1,4 +1,4 @@
-package dojo.rooms.level_4;
+package dojo.rooms.level_2;
 
 import contrib.components.InteractionComponent;
 import contrib.entities.EntityFactory;
@@ -25,7 +25,7 @@ import javax.tools.*;
  * werden muss. Der Dämon erteilt die Aufgabe, Methoden zu schreiben, die dieses Verhalten
  * implementieren. Danach muss der Dämon angegriffen werden.
  */
-public class L4_R3_Monster_Implement_2 extends Room {
+public class L2_R3_Monster_Implement_2 extends Room {
 
   private static final String IMP_FQC = "dojo.monster.MyImp";
   private static final String IMP_PATH = "src/dojo/monster/MyImp";
@@ -56,7 +56,7 @@ public class L4_R3_Monster_Implement_2 extends Room {
    * @param levelSize the size of this room
    * @param designLabel the design label of this room
    */
-  public L4_R3_Monster_Implement_2(
+  public L2_R3_Monster_Implement_2(
       LevelRoom levelRoom,
       RoomGenerator gen,
       Room nextRoom,
