+++ dojo-dungeon/todo-assets/lvl2r2/UML_Klassendiagramm1.png
