+++ dojo-dungeon/todo-assets/lvl3r2/Monster.java
