+++ dojo-dungeon/todo-assets/lvl3r3/ClsToRefactor.java
