+++ dojo-dungeon/todo-assets/lvl4r2/Cuboid.java
